refactor(workflow-designer): extract copy/paste logic into custom hook
Move copy/paste functionality from keyboard-shortcuts.tsx into a dedicated
use-copy-paste-node hook following the same pattern as use-duplicate-node.
This improves code organization and reusability.
