+++ internal-packages/workflow-designer-ui/src/editor/node/index.ts
@@ -1,2 +1,3 @@
 export * from "./node";
+export * from "./use-copy-paste-node";
 export * from "./use-duplicate-node";
